# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Marketplace API",
    "version": "0.1.0",
    "description": "API do marketplace multi-tenant. Rotas por público: /v1/store (comprador), /v1/seller (vendedor), /v1/admin (operador do tenant), /v1/public (integrações) e /v1/platform (staff). Erros no formato RFC 9457. Gerado por `pnpm gen:openapi` — não edite à mão."
  },
  "servers": [
    {
      "url": "https://api.{plataforma}",
      "variables": {
        "plataforma": {
          "default": "localhost:3100"
        }
      }
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "other"
        ],
        "summary": "GET /health",
        "description": "Infraestrutura.",
        "security": [],
        "responses": {
          "2XX": {
            "description": "Sucesso"
          },
          "4XX": {
            "description": "Erro de cliente (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/hooks/identity/clerk": {
      "post": {
        "tags": [
          "other"
        ],
        "summary": "POST /v1/hooks/identity/clerk",
        "description": "Infraestrutura.",
        "security": [],
        "responses": {
          "2XX": {
            "description": "Sucesso"
          },
          "4XX": {
            "description": "Erro de cliente (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/platform/tenants": {
      "get": {
        "tags": [
          "platform"
        ],
        "summary": "GET /v1/platform/tenants",
        "description": "Staff da plataforma — aplicação Clerk Console, sem tenant.",
        "security": [
          {
            "consoleToken": []
          }
        ],
        "responses": {
          "2XX": {
            "description": "Sucesso"
          },
          "4XX": {
            "description": "Erro de cliente (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "platform"
        ],
        "summary": "POST /v1/platform/tenants",
        "description": "Staff da plataforma — aplicação Clerk Console, sem tenant.",
        "security": [
          {
            "consoleToken": []
          }
        ],
        "responses": {
          "2XX": {
            "description": "Sucesso"
          },
          "4XX": {
            "description": "Erro de cliente (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/platform/tenants/{slug}": {
      "get": {
        "tags": [
          "platform"
        ],
        "summary": "GET /v1/platform/tenants/{slug}",
        "description": "Staff da plataforma — aplicação Clerk Console, sem tenant.",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "consoleToken": []
          }
        ],
        "responses": {
          "2XX": {
            "description": "Sucesso"
          },
          "4XX": {
            "description": "Erro de cliente (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/platform/tenants/{tenantId}/status": {
      "patch": {
        "tags": [
          "platform"
        ],
        "summary": "PATCH /v1/platform/tenants/{tenantId}/status",
        "description": "Staff da plataforma — aplicação Clerk Console, sem tenant.",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "consoleToken": []
          }
        ],
        "responses": {
          "2XX": {
            "description": "Sucesso"
          },
          "4XX": {
            "description": "Erro de cliente (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/seller/widgets": {
      "get": {
        "tags": [
          "seller"
        ],
        "summary": "GET /v1/seller/widgets",
        "description": "Painel do vendedor — organização kind=seller na Clerk.",
        "security": [
          {
            "panelToken": []
          }
        ],
        "responses": {
          "2XX": {
            "description": "Sucesso"
          },
          "4XX": {
            "description": "Erro de cliente (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/store/tenant-context": {
      "get": {
        "tags": [
          "store"
        ],
        "summary": "GET /v1/store/tenant-context",
        "description": "Comprador — tenant resolvido pelo host.",
        "security": [],
        "responses": {
          "2XX": {
            "description": "Sucesso"
          },
          "4XX": {
            "description": "Erro de cliente (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/store/widgets": {
      "get": {
        "tags": [
          "store"
        ],
        "summary": "GET /v1/store/widgets",
        "description": "Comprador — tenant resolvido pelo host.",
        "security": [],
        "responses": {
          "2XX": {
            "description": "Sucesso"
          },
          "4XX": {
            "description": "Erro de cliente (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "store"
        ],
        "summary": "POST /v1/store/widgets",
        "description": "Comprador — tenant resolvido pelo host.",
        "security": [],
        "responses": {
          "2XX": {
            "description": "Sucesso"
          },
          "4XX": {
            "description": "Erro de cliente (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "panelToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token de sessão da Clerk (painéis)"
      },
      "consoleToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token da Clerk Console (staff)"
      },
      "apiKey": {
        "type": "apiKey",
        "name": "X-Api-Key",
        "in": "header"
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "code": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "correlation_id": {
            "type": "string"
          }
        }
      }
    }
  }
}